// types/project.ts
export enum ProjectInformationEnum {
  name = 'Allure',
  copyRight = '© 2023 Allure',
  email = 'allure.beauty@gmail.com ',
  address = 'Ho Chi Minh City, Viet Nam ',
  phone = '+84 90 123 4567',
  website = 'https://allure.beauty.com',
  logo = 'allure-logo.png',
  facebook = 'https://www.facebook.com/allurebeauty',
  twitter = 'https://www.twitter.com/allurebeauty',
}